frontend/requirements.txt
streamlit==1.44.0
requests==2.32.3
python-dotenv==1.1.0 